+++ test/test_nacm_protocol.sh
@@ -223,7 +223,7 @@ fi
 
 new "Kill backend"
 # Check if premature kill
-pid=`pgrep -u root -f clixon_backend`
+pid=$(pgrep -u $BUSER -f clixon_backend)
 if [ -z "$pid" ]; then
     err "backend already dead"
 fi