+++ test/test_yang.sh
@@ -286,12 +286,11 @@ fi
 
 new "Kill backend"
 # Check if premature kill
-pid=$(pgrep -u root -f clixon_backend)
+pid=$(pgrep -u $BUSER -f clixon_backend)
 if [ -z "$pid" ]; then
     err "backend already dead"
 fi
 # kill backend
 stop_backend -f "$cfg"
-sudo pkill -u root -f clixon_backend
 
 rm -rf "$dir"